+++ src/test/groovy/de/dfki/mary/voicebuilding/FestvoxPluginFunctionalTest.groovy
@@ -16,6 +16,7 @@ class FestvoxPluginFunctionalTest {
     @BeforeSuite
     void setup() {
         def projectDir = File.createTempDir()
+        new File(projectDir, 'settings.gradle').createNewFile()
 
         gradle = GradleRunner.create()
                 .withProjectDir(projectDir)