+++ src/test/groovy/de/dfki/mary/voicebuilding/LegacyPluginOneShotFunctionalTest.groovy
@@ -14,6 +14,7 @@ class LegacyPluginOneShotFunctionalTest {
     @BeforeSuite
     void setup() {
         def projectDir = File.createTempDir()
+        new File(projectDir, 'settings.gradle').createNewFile()
 
         gradle = GradleRunner.create()
                 .withProjectDir(projectDir)